Fix CI: cross-compile Go binaries on x86 host instead of QEMU
Problem: Building bettercap/pwngrid from Go source inside QEMU ARM
emulation takes 12+ hours, exceeding the 6h GitHub Actions timeout.

Fix: Cross-compile bettercap (CGO for libpcap) and pwngrid on the
x86 runner using ARM cross-compiler, then install pre-built binaries
into rootfs. Reduces Go compilation from 12h to ~5 minutes.

Changes:
- build.yml: Add cross-compile step (Go 1.24.4, armhf/arm64 libs)
- 00-run.sh: NEW host script to copy pre-built binaries to rootfs
- 01-run-chroot.sh: Skip Go SDK install when pre-built binaries exist
- 02-run-chroot.sh: Skip compilation when pre-built, still install caplets
- .gitattributes: Force LF line endings for shell scripts

+++ stage3/03-bettercap-pwngrid/02-run-chroot.sh
@@ -2,43 +2,52 @@
 
 export PATH=$PATH:/usr/local/go/bin:/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin
 
-# install go packages
-for pkg in bettercap pwngrid; do
-    if [ -d "/home/pi/"/$pkg ] ; then
-        echo -e "\e[32m===> Installing $pkg ===\e[0m"
-        if [ $pkg = "pwngrid" ]; then
-            cd "/home/pi/pwngrid"
-            git pull
-            go mod tidy
-            make
-            make install
-        elif [ $pkg = "bettercap" ]; then
-            cd "/home/pi/bettercap"
-            git pull
-            go mod tidy
-            make
-            make install
+# Skip Go compilation if pre-built binaries are already installed
+# (cross-compiled on x86 host to avoid 12+ hour QEMU ARM emulation builds)
+if [ -f /usr/local/bin/bettercap ] && [ -f /usr/local/bin/pwngrid ]; then
+    echo -e "\e[32m=== Pre-built binaries detected, skipping Go compilation ===\e[0m"
+    echo "bettercap: $(ls -la /usr/local/bin/bettercap)"
+    echo "pwngrid:   $(ls -la /usr/local/bin/pwngrid)"
+else
+    # Compile from source (native ARM builds only - takes ~20 min on real hardware)
+    for pkg in bettercap pwngrid; do
+        if [ -d "/home/pi/"/$pkg ] ; then
+            echo -e "\e[32m===> Installing $pkg ===\e[0m"
+            if [ $pkg = "pwngrid" ]; then
+                cd "/home/pi/pwngrid"
+                git pull
+                go mod tidy
+                make
+                make install
+            elif [ $pkg = "bettercap" ]; then
+                cd "/home/pi/bettercap"
+                git pull
+                go mod tidy
+                make
+                make install
+            fi
+        else
+            echo -e "\e[32m===> Installing $pkg ===\e[0m"
+            if [ $pkg = "pwngrid" ]; then
+                cd "/home/pi"
+                git clone https://github.com/jayofelony/pwngrid.git
+                cd "/home/pi/pwngrid"
+                go mod tidy
+                make
+                make install
+            elif [ $pkg = "bettercap" ]; then
+                cd "/home/pi"
+                git clone --recurse-submodules https://github.com/bettercap/bettercap.git
+                cd "/home/pi/bettercap"
+                go mod tidy
+                make
+                make install
+            fi
         fi
-    else
-        echo -e "\e[32m===> Installing $pkg ===\e[0m"
-        if [ $pkg = "pwngrid" ]; then
-            cd "/home/pi"
-            git clone https://github.com/jayofelony/pwngrid.git
-            cd "/home/pi/pwngrid"
-            go mod tidy
-            make
-            make install
-        elif [ $pkg = "bettercap" ]; then
-            cd "/home/pi"
-            git clone --recurse-submodules https://github.com/bettercap/bettercap.git
-            cd "/home/pi/bettercap"
-            go mod tidy
-            make
-            make install
-        fi
-    fi
-done
-# install bettercap caplets
+    done
+fi
+
+# Install bettercap caplets (always needed, regardless of binary source)
 echo -e "\e[32m=== Installing bettercap caplets ===\e[0m"
 cd "/home/pi/"
 git clone https://github.com/jayofelony/caplets.git
